sudhanshu 5526822 Testing and debugging the profile playmate remove and my information playmate remove, made fixes for the detected issues within he multiple account profiles as much as possible.

+++ app/Http/Controllers/Escort/Profile/ProfileInformationController.php
@@ -474,7 +474,11 @@ class ProfileInformationController extends Controller
              */
             $otherEscortProfile = Escort::find($item->playmate_id);
             $otherEscortProfile->playmates()->detach($escortPorfileIdsWithPlaymate);
-            $item->where('playmate_id',$item->playmate_id)->delete();
+
+            /**
+             * Remove records from the history table that matches user and playmate ids.
+             */
+            $item->where(['user_id'=>$item->user_id,'playmate_id'=>$item->playmate_id,])->delete();
         } catch (\Exception $e) {
             $error = true;
             $message = 'Failed to remove playmate: ' . $e->getMessage();
